Add files via upload
Loading all necessary files to run the regression. This represents an my final version of the code when I first made it. However, I should look into changing a few things. First, the data should be scraped, rather than downloaded. Second, the whole model should be refactored so that each section - scraping, cleaning, analyzing, vizualizations - are their own modules that can be run either independently or as a part of a whole model run. Finally, I should implement parallel processing to the part where I calculate the average adjusted R-squared to the results, since this takes quite a while to process (about 20 minutes).